zig: remove bun.FD.int() method

+++ src/StandaloneModuleGraph.zig
@@ -459,7 +459,7 @@ pub const StandaloneModuleGraph = struct {
         // the final 8 bytes in the file are the length of the module graph with padding, excluding the trailer and offsets
         _ = Syscall.write(cloned_executable_fd, std.mem.asBytes(&total_byte_count));
         if (comptime !Environment.isWindows) {
-            _ = bun.C.fchmod(cloned_executable_fd.int(), 0o777);
+            _ = bun.C.fchmod(cloned_executable_fd.cast(), 0o777);
         }
 
         return cloned_executable_fd;
